数码管可以正常显示
数码管可以正常显示
需要将CH452的RSTI脚悬空

+++ BLE-CC254x-1.4.2.2/Components/hal/include/hal_ch452.h
@@ -144,7 +144,7 @@ extern uint8 HalCH452Read( );
 /*
  * write one byte data to ch452
  */
-extern void HalCH452Write( uint8 data);
+extern void HalCH452Write( uint16 data);
 
 
 /*********************************************************************
